feat(automated-rules): implement automated rules (#34)

+++ src/main/java/io/cryostat/rules/Rules.java
@@ -44,15 +44,23 @@ import io.vertx.core.json.JsonObject;
 import jakarta.annotation.security.RolesAllowed;
 import jakarta.inject.Inject;
 import jakarta.transaction.Transactional;
+import jakarta.ws.rs.BadRequestException;
+import jakarta.ws.rs.ClientErrorException;
 import jakarta.ws.rs.Consumes;
 import jakarta.ws.rs.DELETE;
 import jakarta.ws.rs.GET;
+import jakarta.ws.rs.NotFoundException;
 import jakarta.ws.rs.PATCH;
 import jakarta.ws.rs.POST;
 import jakarta.ws.rs.Path;
+import jakarta.ws.rs.core.MediaType;
+import jakarta.ws.rs.core.Response;
 import org.jboss.resteasy.reactive.RestForm;
 import org.jboss.resteasy.reactive.RestPath;
 import org.jboss.resteasy.reactive.RestQuery;
+import org.jboss.resteasy.reactive.RestResponse;
+import org.jboss.resteasy.reactive.RestResponse.ResponseBuilder;
+import org.jboss.resteasy.reactive.RestResponse.Status;
 
 @Path("/api/v2/rules")
 public class Rules {
@@ -61,25 +69,35 @@ public class Rules {
 
     @GET
     @RolesAllowed("read")
-    public V2Response list() {
-        return V2Response.json(Rule.listAll());
+    public RestResponse<V2Response> list() {
+        return RestResponse.ok(V2Response.json(Rule.listAll(), Status.OK.getReasonPhrase()));
     }
 
     @GET
     @RolesAllowed("read")
     @Path("/{name}")
-    public V2Response get(@RestPath String name) {
-        return V2Response.json(Rule.getByName(name));
+    public RestResponse<V2Response> get(@RestPath String name) {
+        return RestResponse.ok(V2Response.json(Rule.getByName(name), Status.OK.getReasonPhrase()));
     }
 
     @Transactional
     @POST
     @RolesAllowed("write")
-    @Consumes("application/json")
-    public V2Response create(Rule rule) {
+    @Consumes({MediaType.APPLICATION_JSON})
+    public RestResponse<V2Response> create(Rule rule) {
         // TODO validate the incoming rule
+        if (rule == null) {
+            throw new BadRequestException("POST body was null");
+        }
+        boolean ruleExists = Rule.getByName(rule.name) != null;
+        if (ruleExists) {
+            throw new RuleExistsException(rule.name);
+        }
         rule.persist();
-        return V2Response.json(rule);
+        return ResponseBuilder.create(
+                        Response.Status.CREATED,
+                        V2Response.json(rule.name, Status.CREATED.toString()))
+                .build();
     }
 
     @Transactional
@@ -87,18 +105,25 @@ public class Rules {
     @RolesAllowed("write")
     @Path("/{name}")
     @Consumes("application/json")
-    public V2Response update(@RestPath String name, @RestQuery boolean clean, JsonObject body) {
+    public RestResponse<V2Response> update(
+            @RestPath String name, @RestQuery boolean clean, JsonObject body) {
         Rule rule = Rule.getByName(name);
-        rule.enabled = body.getBoolean("enabled");
+        boolean enabled = body.getBoolean("enabled");
+        // order matters here, we want to clean before we disable
+        if (clean && !enabled) {
+            bus.send(Rule.RULE_ADDRESS + "?clean", rule);
+        }
+        rule.enabled = enabled;
         rule.persist();
 
-        return V2Response.json(rule);
+        return ResponseBuilder.ok(V2Response.json(rule, Status.OK.toString())).build();
     }
 
     @Transactional
     @POST
     @RolesAllowed("write")
-    public V2Response create(
+    @Consumes({MediaType.MULTIPART_FORM_DATA, MediaType.APPLICATION_FORM_URLENCODED})
+    public RestResponse<V2Response> create(
             @RestForm String name,
             @RestForm String description,
             @RestForm String matchExpression,
@@ -127,9 +152,21 @@ public class Rules {
     @DELETE
     @RolesAllowed("write")
     @Path("/{name}")
-    public V2Response delete(@RestPath String name, @RestQuery boolean clean) {
+    public RestResponse<V2Response> delete(@RestPath String name, @RestQuery boolean clean) {
         Rule rule = Rule.getByName(name);
+        if (rule == null) {
+            throw new NotFoundException("Rule with name " + name + " not found");
+        }
+        if (clean) {
+            bus.send(Rule.RULE_ADDRESS + "?clean", rule);
+        }
         rule.delete();
-        return V2Response.json(rule);
+        return RestResponse.ok(V2Response.json(null, Status.OK.toString()));
+    }
+
+    static class RuleExistsException extends ClientErrorException {
+        RuleExistsException(String ruleName) {
+            super("Rule with name " + ruleName + " already exists", Response.Status.CONFLICT);
+        }
     }
 }
